Move list related methods to extension
